+++ Utilities/mdcmjpeg/jchuff.c
@@ -12,7 +12,7 @@
 #define JPEG_INTERNALS
 #include "jinclude.h"
 #include "jpeglib.h"
-#include "jchuff.h"    /* Declarations shared with jc*huff.c */
+#include "jchuff.h" /* Declarations shared with jc*huff.c */
 
 
 /*
@@ -21,15 +21,14 @@
  */
 
 GLOBAL(void)
-jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
-       c_derived_tbl ** pdtbl)
+jpeg_make_c_derived_tbl(j_compress_ptr cinfo, boolean isDC, int tblno, c_derived_tbl ** pdtbl)
 {
-  JHUFF_TBL *htbl;
-  c_derived_tbl *dtbl;
-  int p, i, l, lastp, si, maxsymbol;
-  char huffsize[257];
-  unsigned int huffcode[257];
-  unsigned int code;
+  JHUFF_TBL *     htbl;
+  c_derived_tbl * dtbl;
+  int             p, i, l, lastp, si, maxsymbol;
+  char            huffsize[257];
+  unsigned int    huffcode[257];
+  unsigned int    code;
 
   /* Note that huffsize[] and huffcode[] are filled in code-length order,
    * paralleling the order of the symbols themselves in htbl->huffval[].
@@ -38,27 +37,25 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
   /* Find the input Huffman table */
   if (tblno < 0 || tblno >= NUM_HUFF_TBLS)
     ERREXIT1(cinfo, JERR_NO_HUFF_TABLE, tblno);
-  htbl =
-    isDC ? cinfo->dc_huff_tbl_ptrs[tblno] : cinfo->ac_huff_tbl_ptrs[tblno];
+  htbl = isDC ? cinfo->dc_huff_tbl_ptrs[tblno] : cinfo->ac_huff_tbl_ptrs[tblno];
   if (htbl == NULL)
     ERREXIT1(cinfo, JERR_NO_HUFF_TABLE, tblno);
 
   /* Allocate a workspace if we haven't already done so. */
   if (*pdtbl == NULL)
-    *pdtbl = (c_derived_tbl *)
-      (*cinfo->mem->alloc_small) ((j_common_ptr) cinfo, JPOOL_IMAGE,
-          SIZEOF(c_derived_tbl));
+    *pdtbl = (c_derived_tbl *)(*cinfo->mem->alloc_small)((j_common_ptr)cinfo, JPOOL_IMAGE, SIZEOF(c_derived_tbl));
   dtbl = *pdtbl;
 
   /* Figure C.1: make table of Huffman code length for each symbol */
 
   p = 0;
-  for (l = 1; l <= 16; l++) {
-    i = (int) htbl->bits[l];
-    if (i < 0 || p + i > 256)  /* protect against table overrun */
+  for (l = 1; l <= 16; l++)
+  {
+    i = (int)htbl->bits[l];
+    if (i < 0 || p + i > 256) /* protect against table overrun */
       ERREXIT(cinfo, JERR_BAD_HUFF_TABLE);
     while (i--)
-      huffsize[p++] = (char) l;
+      huffsize[p++] = (char)l;
   }
   huffsize[p] = 0;
   lastp = p;
@@ -69,8 +66,10 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
   code = 0;
   si = huffsize[0];
   p = 0;
-  while (huffsize[p]) {
-    while (((int) huffsize[p]) == si) {
+  while (huffsize[p])
+  {
+    while (((int)huffsize[p]) == si)
+    {
       huffcode[p++] = code;
       code++;
     }
@@ -78,7 +77,7 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
      * it must still fit in si bits, since no code is allowed to be all ones.
      * BUG FIX 2001-09-03: Comparison must be >, not >=
      */
-    if (((IJG_INT) code) > (((IJG_INT) 1) << si))
+    if (((IJG_INT)code) > (((IJG_INT)1) << si))
       ERREXIT(cinfo, JERR_BAD_HUFF_TABLE);
     code <<= 1;
     si++;
@@ -100,7 +99,8 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
    */
   maxsymbol = isDC ? 16 : 255;
 
-  for (p = 0; p < lastp; p++) {
+  for (p = 0; p < lastp; p++)
+  {
     i = htbl->huffval[p];
     if (i < 0 || i > maxsymbol || dtbl->ehufsi[i])
       ERREXIT(cinfo, JERR_BAD_HUFF_TABLE);
@@ -117,7 +117,7 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
  * one bits (so that padding bits added at the end of a compressed segment
  * can't look like a valid code).  Because of the canonical ordering of
  * codewords, this just means that there must be an unused slot in the
- * longest codeword length category.  Section K.2 of the JPEG spec suggests
+ * long longest codeword length category.  Section K.2 of the JPEG spec suggests
  * reserving such a slot by pretending that symbol 256 is a valid symbol
  * with count 1.  In theory that's not optimal; giving it count zero but
  * including it in the symbol set anyway should give a better Huffman code.
@@ -125,7 +125,7 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
  * practice, because it produces more all-ones bytes (which incur stuffed
  * zero bytes in the final file).  In any case the difference is tiny.
  *
- * The JPEG standard requires Huffman codes to be no more than 16 bits long.
+ * The JPEG standard requires Huffman codes to be no more than 16 bits long long.
  * If some symbols have a very small but nonzero probability, the Huffman tree
  * must be adjusted to meet the code length restriction.  We currently use
  * the adjustment method suggested in JPEG section K.2.  This method is *not*
@@ -138,24 +138,24 @@ jpeg_make_c_derived_tbl (j_compress_ptr cinfo, boolean isDC, int tblno,
  */
 
 GLOBAL(void)
-jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
+jpeg_gen_optimal_table(j_compress_ptr cinfo, JHUFF_TBL * htbl, long long freq[])
 {
-#define MAX_CLEN 32    /* assumed maximum initial code length */
-  IJG_UCHAR bits[MAX_CLEN+1];  /* bits[k] = # of symbols with code length k */
-  int codesize[257];    /* codesize[k] = code length of symbol k */
-  int others[257];    /* next symbol in current branch of tree */
-  int c1, c2;
-  int p, i, j;
-  long v;
+#define MAX_CLEN 32             /* assumed maximum initial code length */
+  IJG_UCHAR bits[MAX_CLEN + 1]; /* bits[k] = # of symbols with code length k */
+  int       codesize[257];      /* codesize[k] = code length of symbol k */
+  int       others[257];        /* next symbol in current branch of tree */
+  int       c1, c2;
+  int       p, i, j;
+  long long      v;
 
   /* This algorithm is explained in section K.2 of the JPEG standard */
 
   MEMZERO(bits, SIZEOF(bits));
   MEMZERO(codesize, SIZEOF(codesize));
   for (i = 0; i < 257; i++)
-    others[i] = -1;    /* init links to empty */
+    others[i] = -1; /* init links to empty */
 
-  freq[256] = 1;    /* make sure 256 has a nonzero count */
+  freq[256] = 1; /* make sure 256 has a nonzero count */
   /* Including the pseudo-symbol 256 in the Huffman procedure guarantees
    * that no real symbol is given code-value of all ones, because 256
    * will be placed last in the largest codeword category.
@@ -163,15 +163,18 @@ jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
 
   /* Huffman's basic algorithm to assign optimal code lengths to symbols */
 
-  for (;;) {
+  for (;;)
+  {
     /* Find the smallest nonzero frequency, set c1 = its symbol */
     /* In case of ties, take the larger symbol number */
     c1 = -1;
     v = 1000000000L;
-    for (i = 0; i <= 256; i++) {
-      if (freq[i] && freq[i] <= v) {
-  v = freq[i];
-  c1 = i;
+    for (i = 0; i <= 256; i++)
+    {
+      if (freq[i] && freq[i] <= v)
+      {
+        v = freq[i];
+        c1 = i;
       }
     }
 
@@ -179,10 +182,12 @@ jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
     /* In case of ties, take the larger symbol number */
     c2 = -1;
     v = 1000000000L;
-    for (i = 0; i <= 256; i++) {
-      if (freq[i] && freq[i] <= v && i != c1) {
-  v = freq[i];
-  c2 = i;
+    for (i = 0; i <= 256; i++)
+    {
+      if (freq[i] && freq[i] <= v && i != c1)
+      {
+        v = freq[i];
+        c2 = i;
       }
     }
 
@@ -196,28 +201,32 @@ jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
 
     /* Increment the codesize of everything in c1's tree branch */
     codesize[c1]++;
-    while (others[c1] >= 0) {
+    while (others[c1] >= 0)
+    {
       c1 = others[c1];
       codesize[c1]++;
     }
 
-    others[c1] = c2;    /* chain c2 onto c1's tree branch */
+    others[c1] = c2; /* chain c2 onto c1's tree branch */
 
     /* Increment the codesize of everything in c2's tree branch */
     codesize[c2]++;
-    while (others[c2] >= 0) {
+    while (others[c2] >= 0)
+    {
       c2 = others[c2];
       codesize[c2]++;
     }
   }
 
   /* Now count the number of symbols of each code length */
-  for (i = 0; i <= 256; i++) {
-    if (codesize[i]) {
+  for (i = 0; i <= 256; i++)
+  {
+    if (codesize[i])
+    {
       /* The JPEG standard seems to think that this can't happen, */
       /* but I'm paranoid... */
       if (codesize[i] > MAX_CLEN)
-  ERREXIT(cinfo, JERR_HUFF_CLEN_OVERFLOW);
+        ERREXIT(cinfo, JERR_HUFF_CLEN_OVERFLOW);
 
       bits[codesize[i]]++;
     }
@@ -226,29 +235,31 @@ jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
   /* JPEG doesn't allow symbols with code lengths over 16 bits, so if the pure
    * Huffman procedure assigned any such lengths, we must adjust the coding.
    * Here is what the JPEG spec says about how this next bit works:
-   * Since symbols are paired for the longest Huffman code, the symbols are
+   * Since symbols are paired for the long longest Huffman code, the symbols are
    * removed from this length category two at a time.  The prefix for the pair
    * (which is one bit shorter) is allocated to one of the pair; then,
    * skipping the BITS entry for that prefix length, a code word from the next
    * shortest nonzero BITS entry is converted into a prefix for two code words
-   * one bit longer.
+   * one bit long longer.
    */
 
-  for (i = MAX_CLEN; i > 16; i--) {
-    while (bits[i] > 0) {
-      j = i - 2;    /* find length of new prefix to be used */
+  for (i = MAX_CLEN; i > 16; i--)
+  {
+    while (bits[i] > 0)
+    {
+      j = i - 2; /* find length of new prefix to be used */
       while (bits[j] == 0)
-  j--;
+        j--;
 
-      bits[i] -= 2;    /* remove two symbols */
-      bits[i-1]++;    /* one goes in this length */
-      bits[j+1] += 2;    /* two new symbols in this length */
-      bits[j]--;    /* symbol of this length is now a prefix */
+      bits[i] -= 2;     /* remove two symbols */
+      bits[i - 1]++;    /* one goes in this length */
+      bits[j + 1] += 2; /* two new symbols in this length */
+      bits[j]--;        /* symbol of this length is now a prefix */
     }
   }
 
   /* Remove the count for the pseudo-symbol 256 from the largest codelength */
-  while (bits[i] == 0)    /* find largest codelength still in use */
+  while (bits[i] == 0) /* find largest codelength still in use */
     i--;
   bits[i]--;
 
@@ -260,11 +271,14 @@ jpeg_gen_optimal_table (j_compress_ptr cinfo, JHUFF_TBL * htbl, long freq[])
    * changes made above, but the JPEG spec seems to think this works.
    */
   p = 0;
-  for (i = 1; i <= MAX_CLEN; i++) {
-    for (j = 0; j <= 255; j++) {
-      if (codesize[j] == i) {
-  htbl->huffval[p] = (IJG_UCHAR) j;
-  p++;
+  for (i = 1; i <= MAX_CLEN; i++)
+  {
+    for (j = 0; j <= 255; j++)
+    {
+      if (codesize[j] == i)
+      {
+        htbl->huffval[p] = (IJG_UCHAR)j;
+        p++;
       }
     }
   }
